pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumos-owasp-defenses"
version = "0.1.0"
description = "OWASP Agentic Security Top 10 defensive implementations"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "AumOS Contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
agentcore = ["agentcore-sdk>=0.1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.3",
    "pip-audit",
]

[project.scripts]
aumos-owasp-defenses = "aumos_owasp_defenses.cli.main:cli"

[project.urls]
Homepage = "https://github.com/aumos-ai/aumos-owasp-defenses"
Documentation = "https://github.com/aumos-ai/aumos-owasp-defenses#readme"
Repository = "https://github.com/aumos-ai/aumos-owasp-defenses"
Issues = "https://github.com/aumos-ai/aumos-owasp-defenses/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/aumos_owasp_defenses"]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=85"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
